新增了头像的上传与下载，把所有ID号全部转为Long类型

+++ GigBookingApp/src/main/java/uk/gigbookingapp/backend/controller/RegisterController.java
@@ -89,7 +89,7 @@ public class RegisterController {
             return Result.error().setMessage("The email has existed.");
         }
         userMapper.insert(user);
-        int id = user.getId();
+        long id = user.getId();
         userPassword.setId(id);
         passwordMapper.insert(userPassword);
